requirements.txt
# framework
django
djangorestframework
# postgresql driver
psycopg2-binary
# swagger
drf-yasg
# to hide data
python-decouple
# filtering
django-filter
# work with images
Pillow
django-ckeditor